feat(project): refactored project code
