+++ .env.example
@@ -1,0 +1,2 @@
+EMAIL=
+PASS=
