--- routes/routesEmailSend.js
@@ -1,8 +1,0 @@
-var express = require("express");
-var app = express();
-var router = express.Router();
-var EmailController = require("../controllers/EmailController");
-
-router.get("/", EmailController.index);
-
-module.exports = router;
