+++ src/index.js
@@ -1,0 +1,13 @@
+const bodyParser = require("body-parser");
+const express = require("express");
+const cors = require("cors");
+const router = require("./routes");
+
+const app = express();
+
+app.use(cors());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+app.use(router);
+
+app.listen(3005, () => console.log("Servidor rodando"));
